requirements.txt
Django>=2.0,<2.1
wagtail>=2.0,<2.1
wagtailfontawesome==1.1.3
django-phonenumber-field==2.0.0

# docs
Sphinx
sphinx_rtd_theme
sphinx-autobuild
